requirements.txt
# aiohttp==3.7.4.post0
# async-timeout==3.0.1
# attrs==21.2.0
beautifulsoup4==4.10.0
# certifi==2021.5.30
# chardet==4.0.0
# charset-normalizer==2.0.6
discord.py==1.7.3
# idna==3.2
# multidict==5.1.0
python-dotenv==0.19.0
# repoze.lru==0.7
# requests==2.26.0
# Routes==2.5.1
# six==1.16.0
# soupsieve==2.2.1
# typing-extensions==3.10.0.2
# urllib3==1.26.7
# yarl==1.6.3
